requirements.txt
pandas==0.25.3
seaborn==0.9.0
jinja2==2.10.3
scikit-learn==0.22.1
numpy==1.18.1
pytorch
scipy
pydotplus
